fix cyclic dependency in scmHistory fyi @lszomoru
